Create groupPrefixes function + remove leading comma from prefixes
